Move createConfirmationDialog method to PointCloudGUI class
